Added title, changed version to be four-parted decimal string. Fixed definition loader not stopping on exit.
git-svn-id: https://jengerer.unfuddle.com/svn/jengerer_jim@176 e479ca42-947a-49ca-8679-5d91e2c49ab0

+++ component.h
@@ -32,8 +32,8 @@ private:
 
 	float x_;
 	float y_;
-	float width_;
-	float height_;
+	int width_;
+	int height_;
 	int alpha_;
 
 };
